chore(val-config-v2): perf optimizations + hardening + cleanup (#2653)
cleanup and hardening pass on the `ValidatorConfigV2` precompile
(stacked on #2635).

- pack `owner` + `init_at_height` into a single `Config` struct (1 SLOAD
per guard instead of 2)
- change `validatorByIndex` type to `uint64` to be consistent with
storage
- strip port from V1 `outboundAddress` during migration — V1 stores
`ip:port` but V2 validates egress as plain IP
- add defense-in-depth check in `migrate_validator` to rejects corrupt
V1 data instead of silently overwriting lookups
- consolidate auth guards + DRY helper fns

+++ tips/ref-impls/src/interfaces/IValidatorConfigV2.sol
@@ -202,7 +202,7 @@ interface IValidatorConfigV2 {
     /// @notice Get validator information by index in the validators array
     /// @param index The index in the validators array
     /// @return The validator struct at the given index
-    function validatorByIndex(uint256 index) external view returns (Validator memory);
+    function validatorByIndex(uint64 index) external view returns (Validator memory);
 
     /// @notice Get validator information by address
     /// @param validatorAddress The validator address to look up
